+++ src/parser/block/extension/placeholder.ts
@@ -33,5 +33,5 @@ export const placeholder: PlaceholderParser = function (source: string): Result<
   }
   const quote = document.createElement('pre');
   void quote.appendChild(document.createTextNode(`${keyword}\n${lines.join('')}${keyword}`));
-  return consumeBlockEndEmptyLine<HTMLElement, SubParsers>([message, quote], source.slice(keyword.length));
+  return consumeBlockEndEmptyLine<HTMLElement, SubParsers>([message, quote], source.slice(keyword.length + 1));
 };