+++ src/parser/segment.ts
@@ -2,7 +2,6 @@
 import { pretext } from '../parser/block/pretext';
 import { extension } from '../parser/block/extension';
 import { combine } from '../combinator/combine';
-import { loop } from '../combinator/loop';
 
 const syntax = /^(?:\s*?\n)+|^(?:[^\n]*\n)+?\s*?\n/;
 assert(!''.match(syntax));
@@ -10,14 +9,11 @@ assert(!' '.match(syntax));
 
 export function segment(source: string): string[] {
   const segments: string[] = [];
-  while (true) {
-    const [, rest] = loop(combine<[PreTextParser, ExtensionParser], HTMLElement>([pretext, extension]))(source) || [[], source.slice((source.match(syntax) || [source])[0].length)];
+  while (source.length > 0) {
+    const [, rest] = combine<[PreTextParser, ExtensionParser], HTMLElement>([pretext, extension])(source) || [[], source.slice((source.match(syntax) || [source])[0].length)];
+    assert(rest.length < source.length);
     void segments.push(source.slice(0, source.length - rest.length));
-    assert(source === '' || source.length > rest.length);
-    source = source.slice(source.length - rest.length);
     source = rest;
-    if (source === '') break;
   }
-  assert(segment.length > 0);
   return segments;
 }